requirements.txt
# Core Framework
fastapi==0.116.1
uvicorn==0.35.0

# AI/ML
google-generativeai==0.8.5

# Data Processing
pandas==2.3.1
numpy==2.3.1

# Visualization
matplotlib>=3.7.0
plotly>=5.15.0
seaborn>=0.12.0

# Configuration
python-dotenv==1.1.1
pydantic==2.11.7
pydantic-settings==2.1.0

# Testing (Development)
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Additional Dependencies
requests==2.31.0